Refactor splits chart-utils into separate modules
